rmagic: Allow float for -w/-h (for svg)
- int works well for png since width/height is measured in pixels, e.g. `-w 640 -h 480`
- But for svg, width/height is measured in some other units (inches?) such that you need to use values more like `-w 8 -h 6`
- This change generalizes width/height from int to float so that svg sizes aren't restricted to int/int, e.g. `-w 10 -h 3.333`

--HG--
branch : version_2.9.x

+++ rpy/ipython/rmagic.py
@@ -508,11 +508,11 @@ utils.install_packages('Cairo')
         )
     @argument_group("Plot", "Arguments to plotting device")
     @argument(
-        '-w', '--width', type=int,
+        '-w', '--width', type=float,
         help='Width of plotting device in R.'
         )
     @argument(
-        '-h', '--height', type=int,
+        '-h', '--height', type=float,
         help='Height of plotting device in R.'
         )
     @argument(
